fixes project and code structure
