+++ src/banking/components/Dashboard.jsx
@@ -1,0 +1,14 @@
+import React from 'react'
+import SidePanel from './SidePanel'
+
+const Dashboard = () => {
+  return (
+    <div className='dash-con'>
+      <SidePanel />
+      <div></div>
+      <div></div>
+    </div>
+  )
+}
+
+export default Dashboard
