+++ src/banking/data/employees.js
@@ -1,0 +1,16 @@
+export const employees = [
+  {
+    username: 'admin',
+    password: 'pass',
+  },
+
+  {
+    username: 'banker001',
+    password: '12345',
+  },
+
+  {
+    username: 'banker002',
+    password: '12345',
+  },
+]
