+++ src/banking/pages/Login.jsx
@@ -1,0 +1,26 @@
+import React from 'react'
+import LoginForm from '../components/LoginForm'
+import '../styles/login.css'
+
+const Login = () => {
+  return (
+    <div className='two-col'>
+      <Logo />
+      <LoginForm />
+    </div>
+  )
+}
+
+const Logo = () => {
+  return (
+    <div className='col'>
+      <img
+        src={require('../../assets/logo.png')}
+        className='logo'
+        alt='logo'
+      ></img>
+    </div>
+  )
+}
+
+export default Login
