+++ src/banking/pages/Main.jsx
@@ -3,11 +3,11 @@ import '../styles/App.css'
 import Dashboard from '../components/Dashboard'
 
 function Main() {
-	return (
-		<div>
-			<Dashboard />
-		</div>
-	)
+  return (
+    <div>
+      <Dashboard />
+    </div>
+  )
 }
 
 export default Main
