+++ src/banking/styles/App.css
@@ -19,7 +19,7 @@ body {
   grid-template-columns: 20% 50% 30%;
 }
 .sideBar {
-   box-shadow: 1px 0px 5px  grey;
+  box-shadow: 1px 0px 5px grey;
 }
 .logo {
   text-align: center;
@@ -35,7 +35,6 @@ body {
 
   transition: all 300ms ease;
   cursor: pointer;
- 
 }
 
 .side-item:hover {
