+++ src/banking/styles/login.css
@@ -1,0 +1,75 @@
+.two-col {
+  width: 100vw;
+  height: 100vh;
+  display: flex;
+}
+
+.col {
+  width: 50%;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.col-two {
+  background-color: #0b2972;
+}
+
+img.logo {
+  width: 50%;
+}
+
+#login-form {
+  width: 50%;
+  padding: 30px;
+  color: white;
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+}
+
+#login-form label {
+  padding-bottom: 10px;
+  color: rgba(255, 255, 255, 0.635);
+}
+
+#login-form input {
+  height: 40px;
+  margin-bottom: 10px;
+  border: none;
+  padding: 0 10px;
+  font-size: 20px;
+  border-radius: 5px;
+}
+
+#login-form input[type='submit'] {
+  width: 100%;
+  margin-top: 20px;
+  color: white;
+  text-shadow: 3px 4px 7px rgba(81, 67, 21, 0.8);
+  background-color: #ffb20c;
+  border: none;
+  cursor: pointer;
+  transition: all 300ms ease;
+}
+
+#login-form input[type='submit']:hover {
+  transform: translateY(-3px);
+}
+
+#login-form h3 {
+  padding-bottom: 30px;
+  text-align: center;
+}
+
+#login-form a,
+#login-form p {
+  font-size: 12px;
+  color: rgba(255, 255, 255, 0.632);
+  text-decoration: none;
+  text-align: center;
+}
+
+#login-form p {
+  padding: 5px 0;
+}
