--- src/components/Dashboard.js
@@ -1,14 +1,0 @@
-import React from 'react'
-import SidePanel from './SidePanel'
-
-const Dashboard = () => {
-    return (
-      <div className='dash-con'>
-          <SidePanel/>
-          <div></div> 
-          <div></div>
-      </div>
-    )
-  }
-
-export default Dashboard
