--- src/saved-data/employees.js
@@ -1,11 +1,0 @@
-export const employees = [
-	{
-		username: 'banker001',
-		password: '12345',
-	},
-
-	{
-		username: 'banker002',
-		password: '12345',
-	},
-]
